+++ extension_packages/figpack_spike_sorting/src/views/view-spike-locations/index.ts
@@ -1,0 +1,3 @@
+export { default as SpikeLocationsView } from "./SpikeLocationsView";
+export { isSpikeLocationsViewData } from "./SpikeLocationsViewData";
+export type { SpikeLocationsViewData } from "./SpikeLocationsViewData";